backend/api-gateway/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pyjwt==2.8.0
